+++ test/support/postgresql_helpers.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-module TestHelper
-  def recreate_table
-    ActiveRecord::Schema.define(version: 0) do
+module PostgresqlTestHelper
+  def recreate_table_in_postgresql
+    PostgresqlDB::Schema.define(version: 0) do
       begin
         drop_table(:things)
       rescue StandardError
@@ -20,17 +20,5 @@ module TestHelper
         t.string :slug
       end
     end
-  end
-
-  def with_migration(&block)
-    migration_class = if ActiveRecord::Migration.respond_to?(:[])
-                        ActiveRecord::Migration[
-                          ActiveRecord::Migration.current_version
-                        ]
-                      else
-                        ActiveRecord::Migration
-                      end
-
-    Class.new(migration_class, &block).new
   end
 end